requirements.txt
# requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
APScheduler==3.10.4
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
Flask-Cors~=5.0.0
alembic~=1.14.0
SQLAlchemy~=2.0.36